A generic load model class.

Class and function definitions for very generic load models that you can create with a single function call (loadModel) or extend to more specific model types (e.g., loadInterp, loadReg2, loadComp)

Slots

fit

A statistical model, fit to the data and wrapped by the loadModel class for additional functionality specific to load models.

pred.format

A string indicating the format of predictions (flux or conc).

metadata

A metadata object describing the load model.

data

The fitting data for the model (fit).

fitting.function

The function used to create or recreate the loadModel, possibly with new fitting data.

y.trans.function

A function to be applied to the y variable before fitting the model.

retrans.function

A function to be applied to the y predictions before returning their values from predictSolute().

See also